pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "google-cloud-storage>=3.1.0",
    "google-genai",
    "gunicorn>=23.0.0",
    "pillow>=11.2.1",
    "psycopg2-binary>=2.9.10",
    "pydub>=0.25.1",
    "python-dotenv>=1.1.0",
    "pycryptodome>=3.20.0",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.41",
    "werkzeug>=3.1.3",
    "agno",
    "googlemaps==4.10.0",
    "chromadb>=0.4.24",
    "sentence-transformers>=2.2.0"
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
]

[project.scripts]
# If you have any command-line scripts, define them here
# "my-script" = "my_package.module:main_function"

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]  # Look in the current directory
include = ["services*", "routes*", "static*", "templates*", "instance*", "*.py"] # Include specified packages and root .py files
exclude = ["venv*", ".git*"] # Exclude venv and .git
